requirements.txt
# Core data processing
pandas==2.1.4
numpy==1.24.3

# Database connectivity
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# ETL and scheduling
schedule==1.2.1
apache-airflow==2.8.0

# HTTP requests and APIs
requests==2.31.0

# Logging and monitoring
coloredlogs==15.0.1

# Configuration and environment
python-dotenv==1.0.0

# Data validation
cerberus==1.3.5

# File handling
openpyxl==3.1.2

# Development and testing
pytest==7.4.3
pytest-cov==4.1.0
